kafkaservice/requirements.txt
aiokafka==0.12.0
kafka-python==2.0.2
asyncio==3.4.3
Flask==3.1.0
waitress==3.0.2
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
protobuf==5.29.3
Werkzeug==3.1.3